+++ maxkb4j-service/maxkb4j-knowledge/src/main/java/com/maxkb4j/knowledge/service/impl/KnowledgeServiceImpl.java
@@ -115,6 +115,7 @@ public class KnowledgeServiceImpl extends ServiceImpl<KnowledgeMapper, Knowledge
         if (CollectionUtils.isEmpty(idList)) {
             return false;
         }
+        dataPermissionSupport.checkManagePermission(AuthTargetType.KNOWLEDGE, idList);
         return deleteKnowledge(idList);
     }
